+++ tests/testthat/test-requests.R
@@ -1,8 +1,7 @@
 
 
 check_test_key <- function(){
-  test_key <- Sys.getenv('BLS_API_TEST_KEY')
-  skip_if(identical(test_key, ""), "'BLS_API_TEST_KEY' not set.")
+  skip_if(!bls_has_key(), "BLS API key is not set.")
 }
 
 test_that('validate_years works',{
@@ -37,7 +36,7 @@ test_that('validate_years works',{
 
 test_that('live get_series_table request tests', {
   check_test_key()
-  api_key <- Sys.getenv('BLS_API_TEST_KEY')
+  api_key <- bls_get_key()
   t1 <- get_series_table('LNS14000002Q', api_key, 1967, 1987)
   years <- as.numeric(dplyr::pull(t1, 'year'))
   expect_equal(min(years), 1967)
@@ -49,7 +48,7 @@ test_that('live get_series_table request tests', {
 
 test_that('live get_series_tables request tests', {
   check_test_key()
-  api_key <- Sys.getenv('BLS_API_TEST_KEY')
+  api_key <- bls_get_key()
   series_ids <- c('LNS14000001','LNS14000002')
   t1 <- get_series_tables(series_ids, api_key, 1968, 1988)
   expect_setequal(names(t1), series_ids)